KaleidoView
this file creates a rectangle of random color and random size.  it also creates 3 other rectangles of the same color and same size.  the random rectangle is in the lower right hand quadrant of the screen.  the other 3 are placed symmetrically around the midpoint of the screen in the lower left, upper left, and upper right quadrants respectively.  this effect is somewhat similar to the design of a Kaleidoscope, on a much simpler scale.